did the intial  endpoint for the hiring companies

+++ backend/src/main/java/com/job_intel/backend/Dtos/HiringCompaniesDto.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -12,7 +14,7 @@ import lombok.Setter;
 public class HiringCompaniesDto {
 
     String companyName;
-    Double jobGrowthRate;
+    BigDecimal jobGrowthRate;
     Long numOfJobs;
 
 }
